+++ falconeri/Cargo.toml
@@ -19,6 +19,5 @@ prettytable-rs = "0.8.0"
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
-serde_yaml = "0.8"
 structopt = "0.2.10"
 uuid = { version = "0.7", features = ["serde", "v4"] }
