+++ falconeri/src/main.rs
@@ -8,9 +8,6 @@ use structopt::StructOpt;
 
 mod cmd;
 mod description;
-mod inputs;
-mod manifest;
-mod pipeline;
 
 /// Command-line options, parsed using `structopt`.
 #[derive(Debug, StructOpt)]
