+++ falconeri_common/Cargo.toml
@@ -13,6 +13,7 @@ common_failures = "0.1.1"
 diesel = { version = "1.3.0", features = ["chrono", "postgres", "r2d2", "serde_json", "uuidv07"] }
 diesel_migrations = "1.3.0"
 failure = "0.1.1"
+handlebars = "1.0.0-beta.4"
 lazy_static = "1.0.2"
 log = "0.4.3"
 r2d2 = "0.8.4"
