+++ falconeri_common/src/lib.rs
@@ -21,7 +21,9 @@ pub use serde_json;
 pub mod connect_via;
 pub mod db;
 pub mod kubernetes;
+pub mod manifest;
 pub mod models;
+pub mod pipeline;
 pub mod rest_api;
 mod schema;
 pub mod secret;
