+++ falconerid/Cargo.toml
@@ -15,3 +15,6 @@ rocket = "0.4.0"
 rocket_contrib = { version = "0.4.0", features = ["uuid"] }
 serde = "1.0"
 serde_derive = "1.0"
+serde_json = "1.0"
+serde_yaml = "0.8"
+
